requirements.txt
ipython==8.29.0
joblib==1.4.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
seaborn==0.13.2
torch==2.8.0
torchvision==0.23.0